Added Retrofit to handle APIs

+++ app/src/main/java/mmalla/android/com/helpabake/ingredient/Ingredient.java
@@ -4,20 +4,20 @@ package mmalla.android.com.helpabake.ingredient;
 import android.os.Parcel;
 import android.os.Parcelable;
 
-public class Ingredient implements Parcelable{
+public class Ingredient implements Parcelable {
 
-    private int quantity;
+    private Double quantity;
     private String measure;
     private String ingredient;
 
-    public Ingredient(int id, String measure, String ingredient) {
+    public Ingredient(Double id, String measure, String ingredient) {
         this.quantity = id;
         this.measure = measure;
         this.ingredient = ingredient;
     }
 
     protected Ingredient(Parcel in) {
-        quantity = in.readInt();
+        quantity = in.readDouble();
         measure = in.readString();
         ingredient = in.readString();
     }
@@ -34,11 +34,11 @@ public class Ingredient implements Parcelable{
         }
     };
 
-    public int getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(Double quantity) {
         this.quantity = quantity;
     }
 
@@ -65,7 +65,7 @@ public class Ingredient implements Parcelable{
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(quantity);
+        dest.writeDouble(quantity);
         dest.writeString(measure);
         dest.writeString(ingredient);
     }
